pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "creditriskengine"
version = "0.1.0"
description = "Production-grade open-source credit risk analytics: Basel III/CRR3 RWA, IFRS 9/CECL ECL, IRB modeling, and model validation"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    {name = "Ankit Jha", email = "ankitjha67@github.com"},
]
keywords = ["credit-risk", "basel-iii", "irb", "ifrs9", "cecl", "rwa", "ecl", "banking", "risk-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pandas>=2.2.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.4.0",
    "statsmodels>=0.14.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
    "hypothesis>=6.98.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=creditriskengine --cov-report=term-missing -v"
